split out tests, increase forkCount
