+++ SingularityService/pom.xml
@@ -466,6 +466,7 @@
           <artifactId>maven-surefire-plugin</artifactId>
           <configuration>
             <forkedProcessTimeoutInSeconds>300</forkedProcessTimeoutInSeconds>
+            <forkCount>2C</forkCount>
           </configuration>
         </plugin>
       </plugins>
