+++ SingularityService/src/test/java/com/hubspot/singularity/scheduler/SingularitySchedulerTest.java
@@ -1,9 +1,7 @@
 package com.hubspot.singularity.scheduler;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -21,7 +19,6 @@ import org.junit.Test;
 import org.mockito.Matchers;
 import org.mockito.Mockito;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
 import com.google.common.base.Optional;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Sets;
@@ -47,12 +44,10 @@ import com.hubspot.singularity.ScheduleType;
 import com.hubspot.singularity.SingularityDeleteResult;
 import com.hubspot.singularity.SingularityDeploy;
 import com.hubspot.singularity.SingularityDeployBuilder;
-import com.hubspot.singularity.SingularityDeployProgress;
 import com.hubspot.singularity.SingularityDeployResult;
 import com.hubspot.singularity.SingularityDeployStatistics;
 import com.hubspot.singularity.SingularityKilledTaskIdRecord;
 import com.hubspot.singularity.SingularityLoadBalancerUpdate;
-import com.hubspot.singularity.SingularityPendingDeploy;
 import com.hubspot.singularity.SingularityPendingRequest;
 import com.hubspot.singularity.SingularityPendingRequest.PendingType;
 import com.hubspot.singularity.SingularityPendingTask;
@@ -65,10 +60,8 @@ import com.hubspot.singularity.SingularityRequestHistory.RequestHistoryType;
 import com.hubspot.singularity.SingularityRequestLbCleanup;
 import com.hubspot.singularity.SingularityTask;
 import com.hubspot.singularity.SingularityTaskHealthcheckResult;
-import com.hubspot.singularity.SingularityTaskHistoryUpdate;
 import com.hubspot.singularity.SingularityTaskId;
 import com.hubspot.singularity.SingularityTaskRequest;
-import com.hubspot.singularity.SingularityUpdatePendingDeployRequest;
 import com.hubspot.singularity.SlavePlacement;
 import com.hubspot.singularity.TaskCleanupType;
 import com.hubspot.singularity.api.SingularityBounceRequest;
@@ -80,13 +73,10 @@ import com.hubspot.singularity.api.SingularityPauseRequest;
 import com.hubspot.singularity.api.SingularityPriorityFreeze;
 import com.hubspot.singularity.api.SingularityRunNowRequest;
 import com.hubspot.singularity.api.SingularityScaleRequest;
-import com.hubspot.singularity.api.SingularitySkipHealthchecksRequest;
 import com.hubspot.singularity.api.SingularityUnpauseRequest;
 import com.hubspot.singularity.data.AbstractMachineManager.StateChangeResult;
 import com.hubspot.singularity.data.SingularityValidator;
-import com.hubspot.singularity.scheduler.SingularityNewTaskChecker.CheckTaskState;
 import com.hubspot.singularity.scheduler.SingularityTaskReconciliation.ReconciliationState;
-import com.sun.jersey.api.ConflictException;
 
 public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   @Inject
@@ -139,653 +129,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     }
 
     Assert.assertTrue(found);
-  }
-
-  @Test
-  public void testDeployManagerHandlesFailedLBTask() {
-    initLoadBalancedRequest();
-    initFirstDeploy();
-
-    SingularityTask firstTask = startTask(firstDeploy);
-
-    initSecondDeploy();
-
-    SingularityTask secondTask = startTask(secondDeploy);
-
-    // this should cause an LB call to happen:
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(taskManager.getLoadBalancerState(secondTask.getTaskId(), LoadBalancerRequestType.ADD).isPresent());
-    Assert.assertTrue(!taskManager.getLoadBalancerState(secondTask.getTaskId(), LoadBalancerRequestType.DEPLOY).isPresent());
-    Assert.assertTrue(!taskManager.getLoadBalancerState(secondTask.getTaskId(), LoadBalancerRequestType.REMOVE).isPresent());
-
-    statusUpdate(secondTask, TaskState.TASK_FAILED);
-    statusUpdate(firstTask, TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState() == DeployState.FAILED);
-
-    List<SingularityPendingTask> pendingTasks = taskManager.getPendingTasks();
-
-    Assert.assertTrue(pendingTasks.size() == 1);
-    Assert.assertTrue(pendingTasks.get(0).getPendingTaskId().getDeployId().equals(firstDeployId));
-  }
-
-  @Test
-  public void testDeployClearsObsoleteScheduledTasks() {
-    initRequest();
-    initFirstDeploy();
-    initSecondDeploy();
-
-    SingularityPendingTaskId taskIdOne = new SingularityPendingTaskId(requestId, firstDeployId, System.currentTimeMillis() + TimeUnit.DAYS.toMillis(3), 1, PendingType.IMMEDIATE, System.currentTimeMillis());
-    SingularityPendingTask taskOne = new SingularityPendingTask(taskIdOne, Optional.<List<String>> absent(), Optional.<String> absent(), Optional.<String> absent(), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<Resources>absent());
-
-    SingularityPendingTaskId taskIdTwo = new SingularityPendingTaskId(requestId, firstDeployId, System.currentTimeMillis() + TimeUnit.DAYS.toMillis(1), 2, PendingType.IMMEDIATE, System.currentTimeMillis());
-    SingularityPendingTask taskTwo = new SingularityPendingTask(taskIdTwo, Optional.<List<String>> absent(), Optional.<String> absent(), Optional.<String> absent(), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<Resources>absent());
-
-    SingularityPendingTaskId taskIdThree = new SingularityPendingTaskId(requestId, secondDeployId, System.currentTimeMillis() + TimeUnit.DAYS.toMillis(3), 1, PendingType.IMMEDIATE, System.currentTimeMillis());
-    SingularityPendingTask taskThree = new SingularityPendingTask(taskIdThree, Optional.<List<String>> absent(), Optional.<String> absent(), Optional.<String> absent(), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<Resources>absent());
-
-    SingularityPendingTaskId taskIdFour = new SingularityPendingTaskId(requestId + "hi", firstDeployId, System.currentTimeMillis() + TimeUnit.DAYS.toMillis(3), 5, PendingType.IMMEDIATE, System.currentTimeMillis());
-    SingularityPendingTask taskFour = new SingularityPendingTask(taskIdFour,Optional.<List<String>> absent(), Optional.<String> absent(), Optional.<String> absent(), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<Resources>absent());
-
-    taskManager.savePendingTask(taskOne);
-    taskManager.savePendingTask(taskTwo);
-    taskManager.savePendingTask(taskThree);
-    taskManager.savePendingTask(taskFour);
-
-    launchTask(request, secondDeploy, 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-
-    List<SingularityPendingTaskId> pendingTaskIds = taskManager.getPendingTaskIds();
-
-    Assert.assertTrue(!pendingTaskIds.contains(taskIdOne));
-    Assert.assertTrue(!pendingTaskIds.contains(taskIdTwo));
-
-    Assert.assertTrue(pendingTaskIds.contains(taskIdThree));
-    Assert.assertTrue(pendingTaskIds.contains(taskIdFour));
-  }
-
-  @Test
-  public void testDeployAllInstancesAtOnce() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-
-    initFirstDeploy();
-
-    launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIds().size());
-    Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(2, taskManager.getCleanupTaskIds().size());
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-  }
-
-  @Test
-  public void testDeployOneInstanceAtATime() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    SingularityTask secondTask = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(1, taskManager.getCleanupTaskIds().size());
-    Assert.assertTrue(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-    SingularityDeployProgress deployProgressStepOne = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(1, deployProgressStepOne.getTargetActiveInstances());
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-
-    deployChecker.checkDeploys();
-
-    SingularityDeployProgress deployProgressStepTwo = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertFalse(deployProgressStepTwo.isStepComplete());
-    Assert.assertEquals(2, deployProgressStepTwo.getTargetActiveInstances());
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testIncrementalDeployCancel() {
-    initRequest();
-
-    // Set up incremental deploy that is partly finished
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-    deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    resourceOffers();
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    resourceOffers();
-    // End in-progress incremental deploy setup
-
-    deployResource.cancelDeploy(requestId, secondDeployId);
-    deployChecker.checkDeploys();
-    Assert.assertEquals(taskManager.getCleanupTasks().get(0).getCleanupType(), TaskCleanupType.INCREMENTAL_DEPLOY_CANCELLED);
-
-    // Incremental deploy task should not be shut down while active deploy is below target instances
-    cleaner.drainCleanupQueue();
-    Assert.assertTrue(taskManager.getKilledTaskIdRecords().isEmpty());
-    Assert.assertEquals(taskManager.getCleanupTasks().get(0).getCleanupType(), TaskCleanupType.INCREMENTAL_DEPLOY_CANCELLED);
-
-    launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    cleaner.drainCleanupQueue();
-    Assert.assertFalse(taskManager.getKilledTaskIdRecords().isEmpty());
-  }
-
-  @Test
-  public void testScaleDownDuringDeploy() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-
-    initFirstDeploy();
-
-    launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    resourceOffers();
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(1)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    Assert.assertEquals(1, taskManager.getCleanupTaskIds().size());
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(2, taskManager.getCleanupTaskIds().size());
-
-    // Extra task from the new deploy should get cleaned up as well
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(3, taskManager.getCleanupTaskIds().size());
-  }
-
-  @Test
-  public void testDeployWithManualStep() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    SingularityTask secondTask = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(1, taskManager.getCleanupTaskIds().size());
-    Assert.assertTrue(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-    SingularityDeployProgress deployProgressStepOne = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(1, deployProgressStepOne.getTargetActiveInstances());
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-
-    deployChecker.checkDeploys();
-
-    // Deploy should not have moved to next step even though instances are launched
-    deployProgressStepOne = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(1, deployProgressStepOne.getTargetActiveInstances());
-
-    // Add the 'ok' to move to the next step
-    deployResource.updatePendingDeploy(new SingularityUpdatePendingDeployRequest(requestId, secondDeployId, 2));
-
-    deployChecker.checkDeploys();
-
-    SingularityDeployProgress deployProgressStepTwo = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertFalse(deployProgressStepTwo.isStepComplete());
-    Assert.assertEquals(2, deployProgressStepTwo.getTargetActiveInstances());
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testDeployMultipleInstancesAtOnce() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(4)).build());
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    SingularityTask secondTask = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-    SingularityTask thirdTask = launchTask(request, firstDeploy, 3, TaskState.TASK_RUNNING);
-    SingularityTask fourthTask = launchTask(request, firstDeploy, 4, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(2), Optional.<Boolean> absent(), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(2, taskManager.getCleanupTaskIds().size());
-    List<SingularityTaskId> cleanupTaskIds = taskManager.getCleanupTaskIds();
-    Assert.assertTrue(cleanupTaskIds.contains(firstTask.getTaskId()) && cleanupTaskIds.contains(secondTask.getTaskId()));
-    SingularityDeployProgress deployProgressStepOne = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(2, deployProgressStepOne.getTargetActiveInstances());
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-    statusUpdate(secondTask, TaskState.TASK_KILLED);
-
-    deployChecker.checkDeploys();
-
-    SingularityDeployProgress deployProgressStepTwo = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertFalse(deployProgressStepTwo.isStepComplete());
-    Assert.assertEquals(4, deployProgressStepTwo.getTargetActiveInstances());
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(4, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(4, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testCancelDeploy() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployResource.cancelDeploy(requestId, secondDeployId);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(taskManager.getCleanupTaskIds().contains(firstNewTaskId));
-    Assert.assertFalse(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-    Assert.assertEquals(DeployState.CANCELED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-
-  }
-
-  @Test
-  public void testDeployFails() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertFalse(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-    Assert.assertEquals(DeployState.FAILED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testDeployFailsForInvalidRequestState() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    initFirstDeploy();
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.of(false), false);
-    requestManager.pause(request, System.currentTimeMillis(), Optional.<String>absent(), Optional.<String>absent());
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.FAILED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testDeployFailsAfterMaxTaskRetries() {
-    initRequest();
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, firstDeployId);
-    db.setMaxTaskRetries(Optional.of(1));
-    SingularityDeploy deploy = initDeploy(db, System.currentTimeMillis());
-
-    deployChecker.checkDeploys();
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, firstDeployId).isPresent());
-
-    SingularityTask task = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(deployManager.getPendingDeploys().get(0).getCurrentDeployState(), DeployState.WAITING);
-
-    SingularityTask taskTryTwo = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(deployManager.getDeployResult(requestId, firstDeployId).get().getDeployState(), DeployState.FAILED);
-  }
-
-  @Test
-  public void testDeploySucceedsWithTaskRetries() {
-    initRequest();
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, firstDeployId);
-    db.setMaxTaskRetries(Optional.of(1));
-    SingularityDeploy deploy = initDeploy(db, System.currentTimeMillis());
-
-    deployChecker.checkDeploys();
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, firstDeployId).isPresent());
-
-    SingularityTask task = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(deployManager.getPendingDeploys().get(0).getCurrentDeployState(), DeployState.WAITING);
-
-    SingularityTask taskTryTwo = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-    Assert.assertEquals(deployManager.getDeployResult(requestId, firstDeployId).get().getDeployState(), DeployState.SUCCEEDED);
-  }
-
-  @Test
-  public void testLbUpdatesAfterEachDeployStep() {
-    initLoadBalancedRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-    initFirstDeploy();
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    SingularityTask secondTask = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean>absent(), Optional.of(1), Optional.<Boolean> absent(), true);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-    SingularityPendingDeploy pendingDeploy = deployManager.getPendingDeploy(requestId).get();
-    Assert.assertEquals(DeployState.WAITING, pendingDeploy.getCurrentDeployState());
-
-    testingLbClient.setNextBaragonRequestState(BaragonRequestState.WAITING);
-
-    deployChecker.checkDeploys();
-    pendingDeploy = deployManager.getPendingDeploy(requestId).get();
-    Assert.assertEquals(DeployState.WAITING, pendingDeploy.getCurrentDeployState());
-
-    testingLbClient.setNextBaragonRequestState(BaragonRequestState.SUCCESS);
-
-    deployChecker.checkDeploys();
-    Assert.assertTrue(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-
-    pendingDeploy = deployManager.getPendingDeploy(requestId).get();
-    SingularityDeployProgress deployProgressStepOne = pendingDeploy.getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(1, deployProgressStepOne.getTargetActiveInstances());
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-
-    deployChecker.checkDeploys();
-
-    pendingDeploy = deployManager.getPendingDeploy(requestId).get();
-    Assert.assertFalse(pendingDeploy.getDeployProgress().get().isStepComplete());
-    Assert.assertEquals(2, pendingDeploy.getDeployProgress().get().getTargetActiveInstances());
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId secondNewTaskId = null;
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId)) {
-      if (taskId.getInstanceNo() == 2) {
-        secondNewTaskId = taskId;
-      }
-    }
-
-    statusUpdate(taskManager.getTask(secondNewTaskId).get(), TaskState.TASK_RUNNING);
-    testingLbClient.setNextBaragonRequestState(BaragonRequestState.WAITING);
-
-    deployChecker.checkDeploys();
-    pendingDeploy = deployManager.getPendingDeploy(requestId).get();
-    Assert.assertEquals(DeployState.WAITING, pendingDeploy.getCurrentDeployState());
-
-    testingLbClient.setNextBaragonRequestState(BaragonRequestState.SUCCESS);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, secondDeployId).get().getDeployState());
-  }
-
-  @Test
-  public void testCanceledDeployTasksStayActiveUntilReplaced() {
-    initRequest();
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-    SingularityTask secondTask = launchTask(request, firstDeploy, 2, TaskState.TASK_RUNNING);
-
-    deploy(secondDeployId, Optional.<Boolean> absent(), Optional.of(1), Optional.<Boolean> absent(), false);
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    Assert.assertEquals(1, taskManager.getPendingTaskIds().size());
-
-    resourceOffers();
-    Assert.assertEquals(1, taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).size());
-
-    SingularityTaskId firstNewTaskId = taskManager.getActiveTaskIdsForDeploy(requestId, secondDeployId).get(0);
-    statusUpdate(taskManager.getTask(firstNewTaskId).get(), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(1, taskManager.getCleanupTaskIds().size());
-    Assert.assertTrue(taskManager.getCleanupTaskIds().contains(firstTask.getTaskId()));
-    SingularityDeployProgress deployProgressStepOne = deployManager.getPendingDeploys().get(0).getDeployProgress().get();
-    Assert.assertTrue(deployProgressStepOne.isStepComplete());
-    Assert.assertEquals(1, deployProgressStepOne.getTargetActiveInstances());
-
-    cleaner.drainCleanupQueue();
-    statusUpdate(firstTask, TaskState.TASK_KILLED);
-
-    deployChecker.checkDeploys();
-    deployResource.cancelDeploy(requestId, secondDeployId);
-    deployChecker.checkDeploys();
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    List<SingularityPendingTaskId> pendingTaskIds = taskManager.getPendingTaskIds();
-    Assert.assertEquals(1, pendingTaskIds.size());
-    Assert.assertEquals(firstDeployId, pendingTaskIds.get(0).getDeployId());
-
-    cleaner.drainCleanupQueue();
-    List<SingularityTaskId> cleanupTaskIds = taskManager.getCleanupTaskIds();
-    Assert.assertEquals(1, cleanupTaskIds.size());
-    Assert.assertEquals(secondDeployId, cleanupTaskIds.get(0).getDeployId());
-
-    resourceOffers();
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, firstDeployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(0, taskManager.getCleanupTaskIds().size());
-    Assert.assertEquals(2, taskManager.getActiveTaskIdsForDeploy(requestId, firstDeployId).size());
-  }
-
-  @Test
-  public void testAfterDeployWaitsForScheduledTaskToFinish() {
-    initScheduledRequest();
-    initFirstDeploy();
-
-    SingularityTask firstTask = launchTask(request, firstDeploy, 1, TaskState.TASK_RUNNING);
-
-    Assert.assertTrue(taskManager.getPendingTasks().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().contains(firstTask.getTaskId()));
-    Assert.assertEquals(1, taskManager.getActiveTaskIds().size());
-    Assert.assertTrue(taskManager.getCleanupTaskIds().isEmpty());
-
-    deploy("nextDeployId");
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    // no second task should be scheduled
-
-    Assert.assertTrue(taskManager.getPendingTasks().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().contains(firstTask.getTaskId()));
-    Assert.assertEquals(1, taskManager.getActiveTaskIds().size());
-    Assert.assertTrue(!taskManager.getCleanupTaskIds().isEmpty());
-
-    statusUpdate(firstTask, TaskState.TASK_FINISHED);
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    cleaner.drainCleanupQueue();
-
-    Assert.assertTrue(!taskManager.getPendingTasks().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getCleanupTaskIds().isEmpty());
-
-    SingularityPendingTaskId pendingTaskId = taskManager.getPendingTaskIds().get(0);
-
-    Assert.assertEquals("nextDeployId", pendingTaskId.getDeployId());
-    Assert.assertEquals(requestId, pendingTaskId.getRequestId());
   }
 
   @Test
@@ -1030,25 +373,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     Assert.assertTrue(!requestResource.getActiveRequests().isEmpty());
     Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.ACTIVE);
-
-    Assert.assertTrue(!taskManager.getPendingTaskIds().isEmpty());
-  }
-
-  @Test
-  public void testScheduledJobLivesThroughDeploy() {
-    initScheduledRequest();
-    initFirstDeploy();
-
-    createAndSchedulePendingTask(firstDeployId);
-
-    Assert.assertTrue(!taskManager.getPendingTaskIds().isEmpty());
-
-    deploy("d2");
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    deployChecker.checkDeploys();
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
 
     Assert.assertTrue(!taskManager.getPendingTaskIds().isEmpty());
   }
@@ -1386,192 +710,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testSlavePlacementSeparate() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)).setSlavePlacement(Optional.of(SlavePlacement.SEPARATE)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1"), createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().size() == 1);
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().size() == 1);
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-
-    eventListener.taskHistoryUpdateEvent(new SingularityTaskHistoryUpdate(taskManager.getActiveTaskIds().get(0), System.currentTimeMillis(), ExtendedTaskState.TASK_CLEANING, Optional.<String>absent(), Optional.<String>absent()));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().size() == 1);
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2")));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 2);
-  }
-
-  @Test
-  public void testSlavePlacementOptimistic() {
-    initRequest();
-    initFirstDeploy();
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1"), createOffer(20, 20000, "slave2", "host2")));
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(3)).setSlavePlacement(Optional.of(SlavePlacement.OPTIMISTIC)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() < 3);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() < 3);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2")));
-
-    eventListener.taskHistoryUpdateEvent(new SingularityTaskHistoryUpdate(taskManager.getActiveTaskIds().get(0), System.currentTimeMillis(), ExtendedTaskState.TASK_CLEANING, Optional.<String>absent(), Optional.<String>absent()));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 3);
-  }
-
-  @Test
-  public void testSlavePlacementOptimisticSingleOffer() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(3)).setSlavePlacement(Optional.of(SlavePlacement.OPTIMISTIC)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1"), createOffer(20, 20000, "slave2", "host2")));
-
-    eventListener.taskHistoryUpdateEvent(new SingularityTaskHistoryUpdate(taskManager.getActiveTaskIds().get(0), System.currentTimeMillis(), ExtendedTaskState.TASK_CLEANING, Optional.<String>absent(), Optional.<String>absent()));
-
-    Assert.assertTrue(taskManager.getPendingTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 3);
-  }
-
-  @Test
-  public void testSlavePlacementGreedy() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(3)).setSlavePlacement(Optional.of(SlavePlacement.GREEDY)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 3);
-  }
-
-  @Test
-  public void testReservedSlaveAttribute() {
-    Map<String, List<String>> reservedAttributes = new HashMap<>();
-    reservedAttributes.put("reservedKey", Arrays.asList("reservedValue1"));
-    configuration.setReserveSlavesWithAttributes(reservedAttributes);
-
-    initRequest();
-    initFirstDeploy();
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("reservedKey", "reservedValue1"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 0);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2", Optional.<String>absent(), ImmutableMap.of("reservedKey", "notAReservedValue"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-  }
-
-  @Test
-  public void testReservedSlaveWithMatchinRequestAttribute() {
-    Map<String, List<String>> reservedAttributes = new HashMap<>();
-    reservedAttributes.put("reservedKey", Arrays.asList("reservedValue1"));
-    configuration.setReserveSlavesWithAttributes(reservedAttributes);
-
-    Map<String, String> reservedAttributesMap = ImmutableMap.of("reservedKey", "reservedValue1");
-    initRequest();
-    initFirstDeploy();
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), reservedAttributesMap)));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 0);
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)).setRequiredSlaveAttributes(Optional.of(reservedAttributesMap)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("reservedKey", "reservedValue1"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-  }
-
-  @Test
-  public void testAllowedSlaveAttributes() {
-    Map<String, List<String>> reservedAttributes = new HashMap<>();
-    reservedAttributes.put("reservedKey", Arrays.asList("reservedValue1"));
-    configuration.setReserveSlavesWithAttributes(reservedAttributes);
-
-    Map<String, String> allowedAttributes = new HashMap<>();
-    allowedAttributes.put("reservedKey", "reservedValue1");
-
-    initRequest();
-    initFirstDeploy();
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("reservedKey", "reservedValue1"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 0);
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)).setAllowedSlaveAttributes(Optional.of(allowedAttributes)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("reservedKey", "reservedValue1"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-  }
-
-  @Test
-  public void testRequiredSlaveAttributesForRequest() {
-    Map<String, String> requiredAttributes = new HashMap<>();
-    requiredAttributes.put("requiredKey", "requiredValue1");
-
-    initRequest();
-    initFirstDeploy();
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)).setRequiredSlaveAttributes(Optional.of(requiredAttributes)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("requiredKey", "notTheRightValue"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2", Optional.<String>absent(), ImmutableMap.of("notTheRightKey", "requiredValue1"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 0);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2", Optional.<String>absent(), requiredAttributes)));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-  }
-
-  @Test
-  public void testMultipleRequiredAttributes() {
-    Map<String, String> requiredAttributes = new HashMap<>();
-    requiredAttributes.put("requiredKey1", "requiredValue1");
-    requiredAttributes.put("requiredKey2", "requiredValue2");
-
-    initRequest();
-    initFirstDeploy();
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(1)).setRequiredSlaveAttributes(Optional.of(requiredAttributes)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String>absent(), ImmutableMap.of("requiredKey1", "requiredValue1"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2", Optional.<String>absent(), ImmutableMap.of("requiredKey1", "requiredValue1", "someotherkey", "someothervalue"))));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 0);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave2", "host2", Optional.<String>absent(), requiredAttributes)));
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 1);
-  }
-
-  @Test
   public void testLBCleanup() {
     initLoadBalancedRequest();
     initFirstDeploy();
@@ -1648,77 +786,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     Assert.assertFalse(taskManager.getLoadBalancerState(taskOne.getTaskId(), LoadBalancerRequestType.ADD).isPresent());
     Assert.assertFalse(taskManager.getLoadBalancerState(taskOne.getTaskId(), LoadBalancerRequestType.REMOVE).isPresent());
     Assert.assertTrue(taskManager.getCleanupTaskIds().contains(taskOne.getTaskId()));
-  }
-
-  @Test
-  public void testUnpauseOnDeploy() {
-    initRequest();
-    initFirstDeploy();
-
-    requestManager.pause(request, System.currentTimeMillis(), Optional.<String>absent(), Optional.<String>absent());
-
-    boolean exception = false;
-
-    try {
-      deploy("d2");
-    } catch (Exception e) {
-      exception = true;
-    }
-
-    Assert.assertTrue(exception);
-
-    deploy("d3", Optional.of(true));
-
-    Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.DEPLOYING_TO_UNPAUSE);
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-    statusUpdate(taskManager.getActiveTasks().get(0), TaskState.TASK_FAILED);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.PAUSED);
-
-    Assert.assertTrue(taskManager.getActiveTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getPendingTaskIds().isEmpty());
-    Assert.assertTrue(requestManager.getPendingRequests().isEmpty());
-
-    deploy("d4", Optional.of(true));
-
-    Assert.assertTrue(requestManager.getRequest(requestId).get().getState() == RequestState.DEPLOYING_TO_UNPAUSE);
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1")));
-
-    statusUpdate(taskManager.getActiveTasks().get(0), TaskState.TASK_RUNNING);
-    deployChecker.checkDeploys();
-
-    RequestState requestState = requestManager.getRequest(requestId).get().getState();
-
-    Assert.assertTrue(requestState == RequestState.ACTIVE);
-  }
-
-  @Test
-  public void testSkipDeployHealthchecks() {
-    initRequest();
-
-    final String deployId = "deploy_test";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setHealthcheckUri(Optional.of("http://uri"));
-    db.setSkipHealthchecksOnDeploy(Optional.of(true));
-
-    SingularityDeploy deploy = initDeploy(db, System.currentTimeMillis());
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
   }
 
   @Test
@@ -1885,44 +952,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testExpiringPause() {
-    initRequest();
-    initFirstDeploy();
-
-    SingularityTask taskOne = startTask(firstDeploy);
-
-    requestResource.pause(requestId, Optional.of(new SingularityPauseRequest(Optional.<Boolean> absent(), Optional.of(1L), Optional.<String> absent(), Optional.<String>absent())));
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(1, taskManager.getKilledTaskIdRecords().size());
-
-    statusUpdate(taskOne, TaskState.TASK_KILLED);
-
-    resourceOffers();
-
-    Assert.assertEquals(0, taskManager.getActiveTaskIds().size());
-    Assert.assertEquals(0, taskManager.getPendingTasks().size());
-    Assert.assertEquals(RequestState.PAUSED, requestManager.getRequest(requestId).get().getState());
-    Assert.assertEquals(requestId, requestManager.getPausedRequests().iterator().next().getRequest().getId());
-
-    try {
-      Thread.sleep(2);
-    } catch (InterruptedException ie){
-    }
-
-    expiringUserActionPoller.runActionOnPoll();
-
-    resourceOffers();
-
-    Assert.assertEquals(1, taskManager.getActiveTaskIds().size());
-    Assert.assertEquals(0, taskManager.getPendingTasks().size());
-
-    Assert.assertEquals(RequestState.ACTIVE, requestManager.getRequest(requestId).get().getState());
-    Assert.assertEquals(requestId, requestManager.getActiveRequests().iterator().next().getRequest().getId());
-  }
-
-  @Test
   public void testBounce() {
     initRequest();
 
@@ -1966,130 +995,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     Assert.assertTrue(taskManager.getCleanupTaskIds().isEmpty());
     Assert.assertTrue(taskManager.getKilledTaskIdRecords().size() == 3);
-  }
-
-  @Test
-  public void testExpiringBounceGoesAway() {
-    initRequest();
-    initFirstDeploy();
-
-    startTask(firstDeploy, 1);
-
-    requestResource.bounce(requestId,
-        Optional.of(new SingularityBounceRequest(Optional.of(false), Optional.<Boolean> absent(), Optional.of(1L), Optional.<String> absent(), Optional.of("msg"))));
-
-    cleaner.drainCleanupQueue();
-    resourceOffers();
-    runLaunchedTasks();
-    cleaner.drainCleanupQueue();
-    killKilledTasks();
-
-    Assert.assertTrue(taskManager.getCleanupTaskIds().isEmpty());
-    Assert.assertEquals(1, taskManager.getActiveTaskIds().size());
-
-    Assert.assertTrue(!requestManager.getExpiringBounce(requestId).isPresent());
-  }
-
-  @Test
-  public void testExpiringNonIncrementalBounce() {
-    initWithTasks(3);
-
-    requestResource.bounce(requestId,
-        Optional.of(new SingularityBounceRequest(Optional.<Boolean> absent(), Optional.<Boolean> absent(), Optional.of(1L), Optional.of("aid"), Optional.<String> absent())));
-
-    Assert.assertTrue(!requestManager.getCleanupRequests().get(0).getMessage().isPresent());
-    Assert.assertEquals("aid", requestManager.getCleanupRequests().get(0).getActionId().get());
-
-    // creates cleanup tasks:
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(1, requestManager.getPendingRequests().size());
-    Assert.assertEquals(0, requestManager.getCleanupRequests().size());
-    Assert.assertEquals(3, taskManager.getCleanupTaskIds().size());
-
-    // should have 1 pending task and 2 launched
-
-    resourceOffersByNumTasks(2);
-
-    Assert.assertEquals(1, taskManager.getPendingTasks().size());
-    Assert.assertEquals(5, taskManager.getActiveTaskIds().size());
-    Assert.assertEquals(3, taskManager.getCleanupTaskIds().size());
-    Assert.assertEquals(0, requestManager.getPendingRequests().size());
-    Assert.assertEquals(0, requestManager.getCleanupRequests().size());
-
-    try {
-      Thread.sleep(1);
-    } catch (InterruptedException ie) {
-    }
-
-    expiringUserActionPoller.runActionOnPoll();
-
-    resourceOffers();
-    cleaner.drainCleanupQueue();
-    killKilledTasks();
-
-    Assert.assertEquals(3, taskManager.getActiveTaskIds().size());
-    Assert.assertEquals(0, taskManager.getPendingTasks().size());
-    Assert.assertEquals(0, taskManager.getCleanupTaskIds().size());
-    Assert.assertEquals(0, requestManager.getPendingRequests().size());
-    Assert.assertEquals(0, requestManager.getCleanupRequests().size());
-  }
-
-  @Test
-  public void testExpiringIncrementalBounce() {
-    initRequest();
-
-    requestResource.scale(requestId, new SingularityScaleRequest(Optional.of(3), Optional.<Long> absent(), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<String>absent()));
-
-    initFirstDeploy();
-
-    startTask(firstDeploy, 1);
-    startTask(firstDeploy, 2);
-    startTask(firstDeploy, 3);
-
-    requestResource.bounce(requestId,
-        Optional.of(new SingularityBounceRequest(Optional.of(true), Optional.<Boolean> absent(), Optional.of(1L), Optional.<String> absent(), Optional.of("msg"))));
-
-    Assert.assertTrue(requestManager.cleanupRequestExists(requestId));
-    Assert.assertEquals("msg", requestManager.getCleanupRequests().get(0).getMessage().get());
-    Assert.assertTrue(requestManager.getCleanupRequests().get(0).getActionId().isPresent());
-
-    String actionId = requestManager.getCleanupRequests().get(0).getActionId().get();
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertTrue(!requestManager.cleanupRequestExists(requestId));
-    Assert.assertTrue(taskManager.getCleanupTaskIds().size() == 3);
-
-    Assert.assertEquals("msg", taskManager.getCleanupTasks().get(0).getMessage().get());
-    Assert.assertEquals(actionId, taskManager.getCleanupTasks().get(0).getActionId().get());
-
-    startTask(firstDeploy, 4);
-//    launchTask(request, firstDeploy, 5, TaskState.TASK_STARTING);
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(1, taskManager.getKilledTaskIdRecords().size());
-    Assert.assertEquals(4, taskManager.getActiveTaskIds().size());
-
-    try {
-      Thread.sleep(2);
-    } catch (InterruptedException ie) {}
-
-    expiringUserActionPoller.runActionOnPoll();
-
-    cleaner.drainCleanupQueue();
-
-    resourceOffers();
-
-    killKilledTasks();
-
-    Assert.assertTrue(!requestManager.getExpiringBounce(requestId).isPresent());
-    Assert.assertTrue(requestManager.getPendingRequests().isEmpty());
-    Assert.assertTrue(taskManager.getPendingTaskIds().isEmpty());
-    Assert.assertTrue(taskManager.getActiveTaskIds().size() == 3);
-    Assert.assertTrue(!requestManager.cleanupRequestExists(requestId));
-    Assert.assertTrue(taskManager.getCleanupTasks().isEmpty());
   }
 
   @Test
@@ -2264,180 +1169,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testBasicSlaveAndRackState() {
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 1, "slave1", "host1", Optional.of("rack1"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 1, "slave2", "host2", Optional.of("rack2"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 1, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(1, slaveManager.getHistory("slave1").size());
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().equals(slaveManager.getHistory("slave1").get(0)));
-
-    sms.slaveLost(driver, SlaveID.newBuilder().setValue("slave1").build());
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 1);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 1);
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.DEAD) == 1);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.DEAD) == 1);
-
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DEAD);
-    Assert.assertTrue(rackManager.getObject("rack1").get().getCurrentState().getState() == MachineState.DEAD);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 1, "slave3", "host3", Optional.of("rack1"))));
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.DEAD) == 1);
-
-    Assert.assertTrue(rackManager.getHistory("rack1").size() == 3);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 1, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 3);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-
-    sms.slaveLost(driver, SlaveID.newBuilder().setValue("slave1").build());
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.DEAD) == 1);
-    Assert.assertTrue(slaveManager.getHistory("slave1").size() == 4);
-
-    sms.slaveLost(driver, SlaveID.newBuilder().setValue("slave1").build());
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.DEAD) == 1);
-    Assert.assertTrue(slaveManager.getHistory("slave1").size() == 4);
-
-    slaveManager.deleteObject("slave1");
-
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.DEAD) == 0);
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(slaveManager.getHistory("slave1").isEmpty());
-  }
-
-  @Test
-  public void testDecommissioning() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)));
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave2", "host2", Optional.of("rack1"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave3", "host3", Optional.of("rack2"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave4", "host4", Optional.of("rack2"))));
-
-
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-
-    Assert.assertTrue(rackManager.getNumObjectsAtState(MachineState.ACTIVE) == 2);
-    Assert.assertTrue(slaveManager.getNumObjectsAtState(MachineState.ACTIVE) == 4);
-
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size() == 1);
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size() == 1);
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave3").get()).isEmpty());
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave4").get()).isEmpty());
-
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ALREADY_AT_STATE, slaveManager.changeState("slave1", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.FAILURE_NOT_FOUND, slaveManager.changeState("slave9231", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user1")));
-
-    Assert.assertEquals(MachineState.STARTING_DECOMMISSION, slaveManager.getObject("slave1").get().getCurrentState().getState());
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getUser().get().equals("user1"));
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(3)));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size() == 1);
-
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DECOMMISSIONING);
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getUser().get().equals("user1"));
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DECOMMISSIONING);
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getUser().get().equals("user1"));
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave4", "host4", Optional.of("rack2"))));
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave3", "host3", Optional.of("rack2"))));
-
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave4").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave3").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-
-    // all task should have moved.
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave4").get()).size() == 1);
-    Assert.assertTrue(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave3").get()).size() == 1);
-    Assert.assertEquals(1, taskManager.getKilledTaskIdRecords().size());
-
-    // kill the task
-    statusUpdate(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).get(0), TaskState.TASK_KILLED);
-
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DECOMMISSIONED);
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getUser().get().equals("user1"));
-
-    // let's DECOMMission rack2
-    Assert.assertEquals(StateChangeResult.SUCCESS, rackManager.changeState("rack2", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user2")));
-
-    // it shouldn't place any on here, since it's DECOMMissioned
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(0, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(0, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-
-    slaveResource.activateSlave("slave1", Optional.<SingularityMachineChangeRequest> absent());
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-
-    Assert.assertTrue(rackManager.getObject("rack2").get().getCurrentState().getState() == MachineState.DECOMMISSIONING);
-
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave2", "host2", Optional.of("rack1"))));
-
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-
-    cleaner.drainCleanupQueue();
-
-    // kill the tasks
-    statusUpdate(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave3").get()).get(0), TaskState.TASK_KILLED);
-
-    Assert.assertTrue(rackManager.getObject("rack2").get().getCurrentState().getState() == MachineState.DECOMMISSIONING);
-
-    statusUpdate(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave4").get()).get(0), TaskState.TASK_KILLED);
-
-    Assert.assertTrue(rackManager.getObject("rack2").get().getCurrentState().getState() == MachineState.DECOMMISSIONED);
-
-  }
-
-  @Test
   public void testTaskOddities() {
     // test unparseable status update
     TaskStatus.Builder bldr = TaskStatus.newBuilder()
@@ -2495,18 +1226,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     resourceOffers();
 
     Assert.assertEquals(4, taskManager.getActiveTaskIds().size());
-  }
-
-  @Test
-  public void testEmptyDecommissioning() {
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user1")));
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-    sms.resourceOffers(driver, Arrays.asList(createOffer(1, 129, "slave1", "host1", Optional.of("rack1"))));
-
-    Assert.assertEquals(MachineState.DECOMMISSIONED, slaveManager.getObject("slave1").get().getCurrentState().getState());
   }
 
   @Test
@@ -2575,169 +1294,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testExpiringScale() {
-    initRequest();
-    initFirstDeploy();
-
-    requestResource.scale(requestId, new SingularityScaleRequest(Optional.of(5), Optional.of(1L), Optional.<Boolean> absent(), Optional.<String> absent(), Optional.<String>absent()));
-
-    try {
-      Thread.sleep(2);
-    } catch (InterruptedException e) {
-
-    }
-
-    expiringUserActionPoller.runActionOnPoll();
-
-    resourceOffers();
-    resourceOffers();
-    resourceOffers();
-    resourceOffers();
-
-    Assert.assertEquals(1, taskManager.getNumActiveTasks());
-  }
-
-  @Test
-  public void testExpiringSkipHealthchecks() {
-    initRequest();
-    initHCDeploy();
-
-    SingularityTask firstTask = startTask(firstDeploy);
-
-    Assert.assertTrue(healthchecker.cancelHealthcheck(firstTask.getTaskId().getId()));
-
-    requestResource.skipHealthchecks(requestId, new SingularitySkipHealthchecksRequest(Optional.of(true), Optional.of(1L), Optional.<String> absent(), Optional.<String>absent()));
-
-    statusUpdate(firstTask, TaskState.TASK_FAILED);
-
-    SingularityTask secondTask = startTask(firstDeploy);
-
-    Assert.assertFalse(healthchecker.cancelHealthcheck(secondTask.getTaskId().getId()));
-
-    statusUpdate(secondTask, TaskState.TASK_FAILED);
-
-    expiringUserActionPoller.runActionOnPoll();
-
-    SingularityTask thirdTask = startTask(firstDeploy);
-
-    Assert.assertTrue(healthchecker.cancelHealthcheck(thirdTask.getTaskId().getId()));
-  }
-
-  @Test
-  public void testSkipHealthchecksEdgeCases() {
-    configuration.setNewTaskCheckerBaseDelaySeconds(0);
-    configuration.setHealthcheckIntervalSeconds(0);
-    configuration.setDeployHealthyBySeconds(0);
-    configuration.setKillAfterTasksDoNotRunDefaultSeconds(100);
-    configuration.setHealthcheckMaxRetries(Optional.of(0));
-    configuration.setCheckNewTasksEverySeconds(1);
-
-    initRequest();
-    initHCDeploy();
-
-    requestResource.skipHealthchecks(requestId, new SingularitySkipHealthchecksRequest(Optional.of(Boolean.TRUE), Optional.<Long> absent(), Optional.<String> absent(), Optional.<String> absent()));
-
-    SingularityTask firstTask = startTask(firstDeploy, 1);
-
-    Assert.assertTrue(!taskManager.getLastHealthcheck(firstTask.getTaskId()).isPresent());
-
-    finishHealthchecks();
-    finishNewTaskChecksAndCleanup();
-
-    Assert.assertEquals(1, taskManager.getNumActiveTasks());
-
-    requestResource.skipHealthchecks(requestId, new SingularitySkipHealthchecksRequest(Optional.of(Boolean.FALSE), Optional.<Long> absent(), Optional.<String> absent(), Optional.<String> absent()));
-
-    // run new task check ONLY.
-    newTaskChecker.enqueueNewTaskCheck(firstTask, requestManager.getRequest(requestId), healthchecker);
-
-    finishNewTaskChecks();
-    finishHealthchecks();
-    finishNewTaskChecksAndCleanup();
-
-    // healthcheck will fail
-    Assert.assertTrue(taskManager.getLastHealthcheck(firstTask.getTaskId()).isPresent());
-    Assert.assertEquals(0, taskManager.getNumActiveTasks());
-  }
-
-  @Test
-  public void testSkipHealthchecksDuringBounce() {
-    initRequest();
-    initHCDeploy();
-
-    SingularityTask firstTask = startTask(firstDeploy, 1);
-
-    requestResource.bounce(requestId, Optional.of(new SingularityBounceRequest(Optional.<Boolean> absent(), Optional.of(true), Optional.<Long> absent(), Optional.<String> absent(), Optional.<String>absent())));
-
-    configuration.setNewTaskCheckerBaseDelaySeconds(0);
-    configuration.setHealthcheckIntervalSeconds(0);
-    configuration.setDeployHealthyBySeconds(0);
-    configuration.setKillAfterTasksDoNotRunDefaultSeconds(1);
-    configuration.setHealthcheckMaxRetries(Optional.of(0));
-
-    cleaner.drainCleanupQueue();
-    resourceOffers();
-
-    List<SingularityTaskId> taskIds = taskManager.getAllTaskIds();
-    taskIds.remove(firstTask.getTaskId());
-
-    SingularityTaskId secondTaskId = taskIds.get(0);
-
-    SingularityTask secondTask = taskManager.getTask(secondTaskId).get();
-
-    statusUpdate(secondTask, TaskState.TASK_RUNNING);
-
-    Assert.assertTrue(healthchecker.cancelHealthcheck(firstTask.getTaskId().getId()));
-
-    newTaskChecker.cancelNewTaskCheck(firstTask.getTaskId().getId());
-
-    finishHealthchecks();
-    finishNewTaskChecks();
-
-    Assert.assertTrue(!taskManager.getLastHealthcheck(secondTask.getTaskId()).isPresent());
-
-    cleaner.drainCleanupQueue();
-    killKilledTasks();
-
-    Assert.assertEquals(0, taskManager.getNumCleanupTasks());
-    Assert.assertEquals(1, taskManager.getNumActiveTasks());
-  }
-
-  @Test
-  public void testHealthchecksDuringBounce() {
-    initRequest();
-    initHCDeploy();
-
-    startTask(firstDeploy);
-
-    requestResource.bounce(requestId);
-
-    cleaner.drainCleanupQueue();
-
-    SingularityTask secondTask = startTask(firstDeploy);
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(1, taskManager.getNumCleanupTasks());
-    Assert.assertEquals(2, taskManager.getNumActiveTasks());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), 1, Optional.<String> absent(), Optional.<String> absent(), secondTask.getTaskId()));
-
-    cleaner.drainCleanupQueue();
-
-    Assert.assertEquals(1, taskManager.getNumCleanupTasks());
-    Assert.assertEquals(2, taskManager.getNumActiveTasks());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(200), Optional.of(1000L), System.currentTimeMillis(), Optional.<String> absent(), Optional.<String> absent(), secondTask.getTaskId()));
-
-    cleaner.drainCleanupQueue();
-    killKilledTasks();
-
-    Assert.assertEquals(0, taskManager.getNumCleanupTasks());
-    Assert.assertEquals(1, taskManager.getNumActiveTasks());
-  }
-
-  @Test
   public void testWaitAfterTaskWorks() {
     initRequest();
     initFirstDeploy();
@@ -2762,140 +1318,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testFrozenSlaveTransitions() {
-    initRequest();
-    initFirstDeploy();
-
-    resourceOffers();
-
-    // test transitions out of frozen
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ALREADY_AT_STATE, slaveManager.changeState("slave1", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ILLEGAL_TRANSITION, slaveManager.changeState("slave1", MachineState.DECOMMISSIONING, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ILLEGAL_TRANSITION, slaveManager.changeState("slave1", MachineState.DECOMMISSIONED, Optional.<String> absent(), Optional.of("user1")));
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.ACTIVE, Optional.<String> absent(), Optional.of("user1")));
-
-    // test transitions into frozen
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave2", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ILLEGAL_TRANSITION, slaveManager.changeState("slave2", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave2", MachineState.DECOMMISSIONING, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ILLEGAL_TRANSITION, slaveManager.changeState("slave2", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave2", MachineState.DECOMMISSIONED, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.FAILURE_ILLEGAL_TRANSITION, slaveManager.changeState("slave2", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave2", MachineState.ACTIVE, Optional.<String> absent(), Optional.of("user2")));
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave2", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user2")));
-  }
-
-  @Test
-  public void testFrozenSlaveDoesntLaunchTasks() {
-    initRequest();
-    initFirstDeploy();
-
-    resourceOffers();
-
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user1")));
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)));
-
-    resourceOffers();
-
-    Assert.assertEquals(0, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-    Assert.assertEquals(2, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size());
-  }
-
-  @Test
-  public void testUnfrozenSlaveLaunchesTasks() {
-    initRequest();
-    initFirstDeploy();
-
-    resourceOffers();
-
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user1")));
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)).setSlavePlacement(Optional.of(SlavePlacement.SEPARATE)));
-
-    resourceOffers();
-
-    Assert.assertEquals(0, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size());
-
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.ACTIVE, Optional.<String> absent(), Optional.of("user1")));
-
-    resourceOffers();
-
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size());
-  }
-
-  @Test
-  public void testFrozenSlaveCanBeDecommissioned() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)));
-
-    resourceOffers();
-
-    // freeze slave1
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.FROZEN, Optional.<String> absent(), Optional.of("user1")));
-
-    // mark tasks as running
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-
-    // assert Request is spread over the two slaves
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-    Assert.assertEquals(1, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size());
-
-    // decommission frozen slave1
-    Assert.assertEquals(StateChangeResult.SUCCESS, slaveManager.changeState("slave1", MachineState.STARTING_DECOMMISSION, Optional.<String> absent(), Optional.of("user1")));
-
-    resourceOffers();
-    cleaner.drainCleanupQueue();
-
-    // assert slave1 is decommissioning
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DECOMMISSIONING);
-
-    // mark tasks as running
-    for (SingularityTask task : taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get())) {
-      statusUpdate(task, TaskState.TASK_RUNNING);
-    }
-
-    // all tasks should have moved
-    cleaner.drainCleanupQueue();
-
-    // kill decommissioned task
-    statusUpdate(taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).get(0), TaskState.TASK_KILLED);
-
-    // assert all tasks on slave2 + slave1 is decommissioned
-    Assert.assertEquals(0, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave1").get()).size());
-    Assert.assertEquals(2, taskManager.getTasksOnSlave(taskManager.getActiveTaskIds(), slaveManager.getObject("slave2").get()).size());
-    Assert.assertTrue(slaveManager.getObject("slave1").get().getCurrentState().getState() == MachineState.DECOMMISSIONED);
-  }
-
-  @Test
-  public void testDeployTimesOut() {
-    initRequest();
-
-    final long hourAgo = System.currentTimeMillis() - TimeUnit.HOURS.toMillis(1);
-
-    final String deployId = "timeout_test";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setDeployHealthTimeoutSeconds(Optional.of(TimeUnit.MINUTES.toSeconds(1)));
-
-    initDeploy(db, hourAgo);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.OVERDUE, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
-  }
-
-  @Test
   public void testRemovedRequestData() {
     initRequest();
     SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, firstDeployId);
@@ -2910,136 +1332,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
     SingularityDeployResult deployResult = deployManager.getDeployResult(requestId, firstDeployId).get();
     Assert.assertEquals(DeployState.FAILED, deployResult.getDeployState());
     Assert.assertTrue(deployResult.getMessage().get().contains("MISSING"));
-  }
-
-  @Test
-  public void testHealthchecksTimeout() {
-    initRequest();
-
-    final long hourAgo = System.currentTimeMillis() - TimeUnit.HOURS.toMillis(1);
-
-    final String deployId = "timeout_test";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setHealthcheckMaxTotalTimeoutSeconds(Optional.of(30L));
-    db.setHealthcheckUri(Optional.of("http://uri"));
-    db.setDeployHealthTimeoutSeconds(Optional.of(TimeUnit.DAYS.toMillis(1)));
-
-    SingularityDeploy deploy = initDeploy(db, hourAgo);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    SingularityTask task = launchTask(request, deploy, hourAgo, hourAgo + 1, 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), hourAgo + 1, Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.FAILED, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
-  }
-
-  @Test
-  public void testMaxHealthcheckRetries() {
-    initRequest();
-
-    final String deployId = "retry_test";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setHealthcheckMaxRetries(Optional.of(2));
-    db.setHealthcheckUri(Optional.of("http://uri"));
-
-    SingularityDeploy deploy = initDeploy(db, System.currentTimeMillis());
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    SingularityTask task = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis(), Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis() + 1, Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis() + 1, Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.FAILED, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
-  }
-
-  @Test
-  public void testNewTaskCheckerRespectsDeployHealthcheckRetries() {
-    initRequest();
-
-    final String deployId = "new_task_healthcheck";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setHealthcheckMaxRetries(Optional.of(1));
-    db.setHealthcheckUri(Optional.of("http://uri"));
-
-    SingularityDeploy deploy = initAndFinishDeploy(request, db);
-
-    SingularityTask task = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_RUNNING);
-
-    Assert.assertEquals(CheckTaskState.CHECK_IF_OVERDUE, newTaskChecker.getTaskState(task, requestManager.getRequest(requestId), healthchecker));
-    Assert.assertTrue(taskManager.getCleanupTaskIds().isEmpty());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis() + 1, Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-
-    Assert.assertEquals(CheckTaskState.CHECK_IF_OVERDUE, newTaskChecker.getTaskState(task, requestManager.getRequest(requestId), healthchecker));
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis() + 1, Optional.<String> absent(), Optional.<String> absent(), task.getTaskId()));
-
-    Assert.assertEquals(CheckTaskState.UNHEALTHY_KILL_TASK, newTaskChecker.getTaskState(task, requestManager.getRequest(requestId), healthchecker));
-  }
-
-  @Test
-  public void testHealthchecksSuccess() {
-    initRequest();
-
-    final String deployId = "hc_test";
-
-    SingularityDeployBuilder db = new SingularityDeployBuilder(requestId, deployId);
-    db.setHealthcheckMaxRetries(Optional.of(2));
-    db.setHealthcheckMaxTotalTimeoutSeconds(Optional.of(30L));
-    db.setHealthcheckUri(Optional.of("http://uri"));
-
-    SingularityDeploy deploy = initDeploy(db, System.currentTimeMillis());
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    SingularityTask task = launchTask(request, deploy, System.currentTimeMillis(), 1, TaskState.TASK_RUNNING);
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(500), Optional.of(1000L), System.currentTimeMillis(), Optional.<String>absent(), Optional.<String>absent(), task.getTaskId()));
-
-    deployChecker.checkDeploys();
-
-    Assert.assertTrue(!deployManager.getDeployResult(requestId, deployId).isPresent());
-
-    taskManager.saveHealthcheckResult(new SingularityTaskHealthcheckResult(Optional.of(200), Optional.of(1000L), System.currentTimeMillis() + 1, Optional.<String>absent(), Optional.<String>absent(), task.getTaskId()));
-
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
   }
 
   @Test
@@ -3375,61 +1667,5 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
 
     // GMT happens first, so EST is a larger timestamp
     Assert.assertEquals(nextRunEST - nextRunGMT, fiveHoursInMilliseconds);
-  }
-
-  @Test
-  public void testUsesNewRequestDataFromPendingDeploy() {
-    initRequest();
-    initFirstDeploy();
-
-    saveAndSchedule(request.toBuilder().setInstances(Optional.of(2)));
-
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    Assert.assertEquals(2, taskManager.getPendingTaskIds().size());
-    Assert.assertEquals(2, requestManager.getRequest(requestId).get().getRequest().getInstancesSafe());
-
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-    SingularityRequest newRequest = request.toBuilder().setInstances(Optional.of(1)).build();
-
-    String deployId = "test_new_request_data";
-    SingularityDeploy deploy = new SingularityDeployBuilder(request.getId(), deployId).setCommand(Optional.of("sleep 100")).build();
-
-    deployResource.deploy(new SingularityDeployRequest(deploy, Optional.<Boolean>absent(), Optional.<String>absent(), Optional.of(newRequest)));
-
-    deployChecker.checkDeploys();
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    List<SingularityPendingTaskId> pendingTaskIdsForNewDeploy = new ArrayList<>();
-    for (SingularityPendingTaskId pendingTaskId : taskManager.getPendingTaskIds()) {
-      if (pendingTaskId.getDeployId().equals(deployId)) {
-        pendingTaskIdsForNewDeploy.add(pendingTaskId);
-      }
-    }
-
-    Assert.assertEquals(1, pendingTaskIdsForNewDeploy.size());
-    Assert.assertEquals(2, requestManager.getRequest(requestId).get().getRequest().getInstancesSafe());
-
-    resourceOffers();
-    for (SingularityTaskId taskId : taskManager.getActiveTaskIdsForDeploy(requestId, deployId)) {
-      statusUpdate(taskManager.getTask(taskId).get(), TaskState.TASK_RUNNING);
-    }
-    deployChecker.checkDeploys();
-
-    Assert.assertEquals(1, requestManager.getRequest(requestId).get().getRequest().getInstancesSafe());
-  }
-
-  @Test(expected = ConflictException.class)
-  public void testCannotUpdateRequestDuringPendingDeployWithNewData() {
-    initRequest();
-    SingularityRequest request = requestResource.getRequest(requestId).getRequest();
-    SingularityRequest newRequest = request.toBuilder().setInstances(Optional.of(1)).build();
-
-    String deployId = "test_new_request_data";
-    SingularityDeploy deploy = new SingularityDeployBuilder(request.getId(), deployId).setCommand(Optional.of("sleep 100")).build();
-
-    deployResource.deploy(new SingularityDeployRequest(deploy, Optional.<Boolean>absent(), Optional.<String>absent(), Optional.of(newRequest)));
-
-    requestResource.postRequest(newRequest);
   }
 }
